smashed some nasty bugs related to relative jump added auxilary carry support to add_to_a MISC widow now prints PSW register and fixed some more nasty djnz bugs

+++ definitions.h
@@ -42,5 +42,5 @@ extern const struct instr instructions[256];
 
 typedef bool bit;
 typedef unsigned char byte;
-typedef short int word;
+typedef int word;
 #endif
